update project 3

+++ # project 6 useing if.py
@@ -1,7 +1,7 @@
 # project 6 useing if
 
 print ("Welcom my student")
-a = int (input ("what is yours degree?\n"))
+a = float (input ("what is yours degree?\n"))
 if a >= 90 :
   print ("excellent")
 elif a >= 75:
@@ -10,3 +10,5 @@ elif a >= 50:
   print (" acceptable ")
 else:
   print ("weak, sudy hard next time")
+
+  #we can use a lot numbers of "elif"
